refactor: move dtos de pasta incorreta
