+++ backend/src/main/java/com/branches/response/ClienteGetResponse.java
@@ -1,4 +1,4 @@
-package com.branches.controller;
+package com.branches.response;
 
 import com.branches.model.TipoCliente;
 import lombok.Builder;
